feat: move UserFeedViewModel, StringFeedViewModel, ServerName to commons (batch 2)
Move the following to the KMP commons module for cross-platform reuse:

- ServerName data class + ServerType enum + DEFAULT_MEDIA_SERVERS
- UserFeedState sealed class + UserFeedViewModel
- StringFeedState sealed class + StringFeedViewModel

ViewModels parameterized with ICacheProvider instead of LocalCache singleton.
Typealias re-exports in original packages for backward compatibility.
Android subclasses updated to pass LocalCache as cacheProvider.

Build verified: :commons:compileKotlinJvm + :amethyst:compilePlayDebugKotlin ✅
