+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/settings/dal/HiddenWordsFeedViewModel.kt
@@ -23,12 +23,14 @@ package com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.dal
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.vitorpamplona.amethyst.model.Account
+import com.vitorpamplona.amethyst.model.LocalCache
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.settings.StringFeedViewModel
 
 class HiddenWordsFeedViewModel(
     val account: Account,
 ) : StringFeedViewModel(
         HiddenWordsFeedFilter(account),
+        LocalCache,
     ) {
     class Factory(
         val account: Account,
